Move is_deployed_environment function to recordtransfer.utils
