+++ app/recordtransfer/views/pre_submission.py
@@ -23,7 +23,6 @@ from django.http import (
     Http404,
     HttpRequest,
     HttpResponse,
-    HttpResponseRedirect,
     QueryDict,
 )
 from django.shortcuts import redirect
@@ -49,13 +48,13 @@ from recordtransfer.emails import (
     send_your_submission_did_not_go_through,
 )
 from recordtransfer.enums import SubmissionStep
-from recordtransfer.management.commands.verify_settings import is_deployed_environment
 from recordtransfer.models import (
     InProgressSubmission,
     Submission,
     UploadSession,
     User,
 )
+from recordtransfer.utils import is_deployed_environment
 
 LOGGER = logging.getLogger(__name__)
